+++ CHANGELOG.md
@@ -9,6 +9,7 @@
  - Convergence errors in models that use Newton-Rhapson methods now throw a `ConvergenceError`, instead of a `ValueError` (the former is a subclass of the latter, however).
  - `AalenAdditiveModel` raises `ConvergenceWarning` instead of printing a warning.
  - `KaplanMeierFitter` now has a cumulative plot option. Example `kmf.plot(invert_y_axis=True)`
+ - a `weights_col` option has been added to `CoxTimeVaryingFitter` that allows for time-varying weights. 
 
 #### 0.14.6
  - fix for n > 2 groups in `multivariate_logrank_test` (again).